+++ webpack.config.dev.js
@@ -1,46 +1,49 @@
-var path = require('path');
-var webpack = require('webpack');
+const path = require("path");
+const webpack = require("webpack");
 
 module.exports = {
-  devtool: 'cheap-module-eval-source-map',
+  devtool: "cheap-module-eval-source-map",
   entry: [
-    'eventsource-polyfill', // necessary for hot reloading with IE
-    'webpack-hot-middleware/client',
-    './src/index',
+    "eventsource-polyfill", // necessary for hot reloading with IE
+    "webpack-hot-middleware/client",
+    "./demo/index",
   ],
   output: {
-    path: path.join(__dirname, 'build'),
-    filename: 'bundle.js',
-    publicPath: '/',
+    path: path.join(__dirname, "build"),
+    filename: "bundle.js",
+    publicPath: "/",
   },
   plugins: [
     new webpack.HotModuleReplacementPlugin(),
     new webpack.NoErrorsPlugin(),
   ],
   resolve: {
-    extensions: ['', '.js', '.jsx'],
+    extensions: ["", ".js", ".jsx"],
   },
   module: {
     loaders: [{
       test: /\.jsx?/,
-      loaders: ['babel', 'strict'],
-      include: path.join(__dirname, 'src'),
+      loaders: ["babel", "strict"],
+      include: [
+        path.join(__dirname, "demo"),
+        path.join(__dirname, "src"),
+      ],
     },
     {
       test: /\.css$/,
-      loader: 'style-loader!css-loader!postcss-loader',
+      loader: "style-loader!css-loader!postcss-loader",
     },
     {
       test: /\.(png|jpg|gif|svg|woff|woff2|ttf|eot)$/,
-      loader: 'url-loader?limit=1',
+      loader: "url-loader?limit=1",
     },
     {
       test: /\.json$/,
-      loader: 'json-loader',
+      loader: "json-loader",
     },
     {
       test: /\.html$/,
-      loader: 'html-loader',
+      loader: "html-loader",
     },
     ],
   },
